# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API for Mobile Application Design & Development (MADD)",
    "description": "RESTful API Documentation with Swagger",
    "contact": {
      "name": "Gerald C. Hurdle",
      "email": "hurdleg@algonquincollege.com",
      "url": "http://algonquinhosting.ca/~hurdleg"
    }
  },
  "host": "madd.mybluemix.net",
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "tags": [
    {
      "name": "Courses",
      "description": "Everything about Courses"
    }
  ],
  "paths": {
    "/courses": {
      "get": {
        "summary": "Get all the Courses",
        "description": "Returns the entire collection of `Course` objects in a JSON array.",
        "operationId": "getCourses",
        "tags": [
          "Courses"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Course"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Add a new Course",
        "description": "",
        "operationId": "postCourse",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Course object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/courses/count": {
      "get": {
        "summary": "Get the number of Courses",
        "description": "Returns the number of courses (i.e. `count`).\n",
        "operationId": "getCount",
        "tags": [
          "Courses"
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "examples": {
              "application/json": {
                "count": 4
              }
            }
          }
        }
      }
    },
    "/courses/{courseId}": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "Find a course by Id",
        "description": "Returns a single course",
        "operationId": "getCourseById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "Id of course to return",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "400": {
            "description": "Invalid Id supplied"
          },
          "404": {
            "description": "Course not found"
          }
        }
      },
      "put": {
        "tags": [
          "Courses"
        ],
        "summary": "Update an existing course",
        "description": "",
        "operationId": "updateCourse",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "Course Id to update",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Course object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "delete": {
        "tags": [
          "Courses"
        ],
        "summary": "Delete a course by Id",
        "description": "",
        "operationId": "deleteCourse",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "Course Id to delete",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid Id supplied"
          },
          "404": {
            "description": "Course not found"
          }
        }
      }
    },
    "/courses/form": {
      "post": {
        "tags": [
          "Courses"
        ],
        "summary": "Add a new course with form data",
        "description": "",
        "operationId": "newCourseForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "formData",
            "description": "New course code. Example, MAD9132",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "New course name. Example, Android Application Development",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "New course description. Example, Let's write native Android apps with Java!",
            "required": true,
            "type": "string"
          },
          {
            "name": "level",
            "in": "formData",
            "description": "New course level. Must be in the range [1..4]. Example, 3",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/courses/{courseId}/form": {
      "put": {
        "tags": [
          "Courses"
        ],
        "summary": "Update an existing course with form data",
        "description": "",
        "operationId": "updateCourseForm",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "Course Id to update",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "code",
            "in": "formData",
            "description": "Updated course code. Example, MAD9145",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Updated course name. Example, Applied Project",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "Updated course description. Example, It's Go Time!",
            "required": true,
            "type": "string"
          },
          {
            "name": "level",
            "in": "formData",
            "description": "Updated course level. Must be in the range [1..4]. Example, 4",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
    "Course": {
      "title": "Course",
      "type": "object",
      "properties": {
        "courseId": {
          "type": "integer",
          "description": "uniquely identifies the course (i.e. PK)"
        },
        "code": {
          "type": "string",
          "description": "the code of this course. For example, MAD9132"
        },
        "name": {
          "type": "string",
          "description": "the name of this course. For example, Android Application Development"
        },
        "description": {
          "type": "string",
          "description": "more information about this course"
        },
        "level": {
          "type": "integer",
          "description": "the level of this course",
          "enum": [
            1,
            2,
            3,
            4
          ]
        }
      }
    }
  }
}